student item delete feature added

+++ 16_SQLite_Database/app/src/main/java/com/hellohasan/sqlite/data/model/student/StudentModel.kt
@@ -6,5 +6,5 @@ interface StudentModel {
     fun insertStudent(student: Student, callback: DataFetchCallback<Student>)
     fun getStudentList(callback: DataFetchCallback<MutableList<Student>>)
     fun updateStudent(student: Student, callback: DataFetchCallback<Student>)
-    fun deleteStudent(id: Long, callback: DataFetchCallback<Long>)
+    fun deleteStudent(id: Long, callback: DataFetchCallback<Int>)
 }
